moving pdf actions
